+++ src/main/java/fi/dy/masa/tweakeroo/tweaks/RenderTweaks.java
@@ -98,6 +98,11 @@ public class RenderTweaks
         fakeWorld = new FakeWorld(dynamicRegistryManager, loadDistance);
     }
 
+    public static void reloadRegistryManager(World world)
+    {
+        dynamicRegistryManager = world.getRegistryManager().toImmutable();
+    }
+
     public static FakeWorld getFakeWorld()
     {
         return fakeWorld;